requirements.txt
# Core dependencies
pillow>=9.0.0
pyautogui>=0.9.54
opencv-python>=4.5.0
numpy>=1.19.0
psutil>=5.8.0

# OCR dependencies (optional)
easyocr>=1.7.0
pytesseract>=0.3.8

# AI/ML dependencies
openai>=1.0.0
requests>=2.25.0

# GUI dependencies
tkinter  # Usually comes with Python

# Utility dependencies
pyperclip>=1.8.2
asyncio>=3.4.3
logging>=0.5.1.2
configparser
json
time
hashlib
pathlib
datetime
typing>=3.7.4.3

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0

# Additional dependencies
python-dotenv>=0.19.0
pywin32>=228; platform_system == "Windows"
keyboard>=0.13.5
mouse>=0.7.1
playwright>=1.30.0
python-json-logger>=2.0.0
aiohttp>=3.8.0 